api-python/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
numpy==2.1.1
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
scikit-learn==1.5.2
scipy==1.14.1
threadpoolctl==3.5.0
tqdm==4.66.5
Werkzeug==3.0.4
